feat: re-organise election mocks and tests
