+++ state-chain/pallets/cf-elections/src/electoral_systems.rs
@@ -6,3 +6,8 @@ pub mod egress_success;
 pub mod mock;
 pub mod monotonic_median;
 pub mod unsafe_median;
+
+#[cfg(test)]
+pub(crate) mod mocks;
+#[cfg(test)]
+pub(crate) mod tests;
