+++ state-chain/pallets/cf-elections/src/electoral_systems/tests.rs
@@ -1,0 +1,7 @@
+pub(crate) use super::mocks;
+
+pub(crate) use crate::{checks, register_checks};
+
+pub mod change;
+pub mod monotonic_median;
+pub mod unsafe_median;
